# docs/swagger.yaml
definitions:
  dto.Chat:
    properties:
      id:
        type: string
      name:
        type: string
    required:
    - id
    type: object
  dto.ChatCreate:
    properties:
      employee_ids:
        items:
          type: string
        type: array
      is_group:
        type: boolean
      name:
        type: string
    type: object
  dto.ChatDetail:
    properties:
      employee_ids:
        items:
          type: string
        type: array
      id:
        type: string
      is_group:
        type: boolean
      name:
        type: string
    required:
    - id
    type: object
  dto.ChatUpdate:
    properties:
      employee_ids:
        items:
          type: string
        type: array
      name:
        type: string
    type: object
  dto.Error:
    properties:
      description:
        type: string
      stackTrace:
        type: string
      status:
        type: integer
    required:
    - description
    - status
    type: object
  dto.Message:
    properties:
      chat_id:
        type: string
      colleague_id:
        type: string
      created_at:
        type: string
      employee_id:
        type: string
      id:
        type: string
      text:
        type: string
    required:
    - chat_id
    - colleague_id
    - created_at
    - employee_id
    - id
    - text
    type: object
  dto.MessageCreate:
    properties:
      chat_id:
        type: string
      colleague_id:
        type: string
      employee_id:
        type: string
      text:
        type: string
    required:
    - chat_id
    - colleague_id
    - employee_id
    - text
    type: object
  dto.MessageUpdate:
    properties:
      text:
        type: string
    required:
    - text
    type: object
info:
  contact: {}
paths:
  /api/v1/chats:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Chat'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Error'
      tags:
      - Chats
    post:
      consumes:
      - application/json
      parameters:
      - description: Chat info
        in: body
        name: chat
        required: true
        schema:
          $ref: '#/definitions/dto.ChatCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Error'
      tags:
      - Chats
  /api/v1/chats/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
      tags:
      - Chats
    get:
      consumes:
      - application/json
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ChatDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
      tags:
      - Chats
    put:
      consumes:
      - application/json
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      - description: Chat info
        in: body
        name: chat
        required: true
        schema:
          $ref: '#/definitions/dto.ChatUpdate'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Error'
      tags:
      - Chats
  /api/v1/chats/user/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Chat'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Error'
      tags:
      - Chats
  /api/v1/messages:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Message'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Error'
      tags:
      - Messages
    post:
      consumes:
      - application/json
      parameters:
      - description: Message info
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/dto.MessageCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Error'
      tags:
      - Messages
  /api/v1/messages/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
      tags:
      - Messages
    get:
      consumes:
      - application/json
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
      tags:
      - Messages
    put:
      consumes:
      - application/json
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      - description: Message info
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/dto.MessageUpdate'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Error'
      tags:
      - Messages
swagger: "2.0"